+++ src/test/java/com/styra/opa/springboot/autoconfigure/properties/DefaultOPAPropertiesTest.java
@@ -1,6 +1,6 @@
-package com.styra.opa.springboot.properties;
+package com.styra.opa.springboot.autoconfigure.properties;
 
-import com.styra.opa.springboot.OPAProperties;
+import com.styra.opa.springboot.autoconfigure.OPAProperties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-@EnableConfigurationProperties(value = OPAProperties.class)
+@EnableConfigurationProperties(OPAProperties.class)
 @ExtendWith(SpringExtension.class)
 public class DefaultOPAPropertiesTest {
 
@@ -22,12 +22,13 @@ public class DefaultOPAPropertiesTest {
     public void test() {
         assertEquals(OPAProperties.DEFAULT_URL, opaProperties.getUrl());
         assertNull(opaProperties.getPath());
-        assertEquals(OPAProperties.DEFAULT_REASON_KEY, opaProperties.getReasonKey());
         assertNotNull(opaProperties.getRequest());
         assertEquals(OPAProperties.Request.Resource.DEFAULT_TYPE, opaProperties.getRequest().getResource().getType());
         assertNotNull(opaProperties.getRequest().getContext());
         assertEquals(OPAProperties.Request.Context.DEFAULT_TYPE, opaProperties.getRequest().getContext().getType());
         assertEquals(OPAProperties.Request.Subject.DEFAULT_TYPE,
                 opaProperties.getRequest().getSubject().getType());
+        assertEquals(OPAProperties.Response.Context.DEFAULT_REASON_KEY,
+                opaProperties.getResponse().getContext().getReasonKey());
     }
 }