+++ src/test/java/com/styra/opa/springboot/autoconfigure/properties/ModifiedSystemEnvOPAPropertiesTest.java
@@ -1,6 +1,6 @@
-package com.styra.opa.springboot.properties;
+package com.styra.opa.springboot.autoconfigure.properties;
 
-import com.styra.opa.springboot.OPAProperties;
+import com.styra.opa.springboot.autoconfigure.OPAProperties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +10,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 // Run using: ./gradlew testModifiedSystemEnvProperties
-@EnableConfigurationProperties(value = OPAProperties.class)
+@EnableConfigurationProperties(OPAProperties.class)
 @ExtendWith(SpringExtension.class)
 public class ModifiedSystemEnvOPAPropertiesTest {
 
@@ -20,6 +20,6 @@ public class ModifiedSystemEnvOPAPropertiesTest {
     @Test
     public void test() {
         assertEquals("http://localhost:8183", opaProperties.getUrl());
-        assertEquals("tickets/main2", opaProperties.getPath());
+        assertEquals("foo/bar2", opaProperties.getPath());
     }
 }